Split out library fairly naively.
