+++ uavcan/src/types.rs
@@ -1,0 +1,10 @@
+//! # Placeholder module for various types I'm not sure how to sort yet.
+
+// Not too sure what to do with these. This is the best option to keep them all seperate.
+pub type NodeId = u8;
+pub type PortId = u16;
+// TODO set type with min/max bounds
+pub type TransferId = u8;
+
+// TODO whole timing overhaul
+pub type Timestamp = std::time::Instant;
